update AutoComplete style.
