+++ demo/AutoComplete.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8"/>
     <title>Carousel - Demo</title>
-    <link rel="stylesheet" href="assets/themes/standard.less" />
+    <link rel="stylesheet" href="assets/themes/standard.css" />
     <link href="assets/third-party/prism.css" rel="stylesheet" />
     <script src="assets/loader/esl.js"></script>
