+++ demo/assets/themes/standard.less
@@ -1,7 +1,7 @@
 @import "../../../dep/est/1.3.0/src/all.less";
 .global-reset();
 
-@import "../../../dep/esf/1.0.0-alpha.4/themes/standard/main-esui.less";
+@import "../../../dep/esf/1.0.0-alpha.7/themes/standard/main-esui.less";
 @import "../../../dep/esui/3.1.0-beta.5/themes/standard/main.less";
 @import "../../../dep/eicons/1.0.0-alpha.1/src/main.less";
